refactor: Split pools hooks
