+++ src/state/hooks.ts
@@ -9,17 +9,9 @@ import { Team } from 'config/constants/types'
 import Nfts from 'config/constants/nfts'
 import { simpleRpcProvider } from 'utils/providers'
 import useRefresh from 'hooks/useRefresh'
-import {
-  fetchPoolsPublicDataAsync,
-  fetchPoolsUserDataAsync,
-  fetchCakeVaultPublicData,
-  fetchCakeVaultUserData,
-  fetchCakeVaultFees,
-  setBlock,
-} from './actions'
+import { setBlock } from './actions'
 import {
   State,
-  Pool,
   ProfileState,
   TeamsState,
   AchievementState,
@@ -33,8 +25,6 @@ import { fetchTeam, fetchTeams } from './teams'
 import { fetchAchievements } from './achievements'
 import { fetchWalletNfts } from './collectibles'
 import { parseBigNumberObj } from './predictions/helpers'
-import { transformPool } from './pools/helpers'
-import { fetchPoolsStakingLimitsAsync } from './pools'
 import {
   fetchCurrentLotteryId,
   fetchCurrentLottery,
@@ -55,129 +45,6 @@ export const usePollBlockNumber = () => {
 
     return () => clearInterval(interval)
   }, [dispatch])
-}
-
-// Pools
-
-export const useFetchPublicPoolsData = () => {
-  const dispatch = useAppDispatch()
-  const { slowRefresh } = useRefresh()
-
-  useEffect(() => {
-    const fetchPoolsPublicData = async () => {
-      const blockNumber = await simpleRpcProvider.getBlockNumber()
-      dispatch(fetchPoolsPublicDataAsync(blockNumber))
-    }
-
-    fetchPoolsPublicData()
-    dispatch(fetchPoolsStakingLimitsAsync())
-  }, [dispatch, slowRefresh])
-}
-
-export const usePools = (account): { pools: Pool[]; userDataLoaded: boolean } => {
-  const { fastRefresh } = useRefresh()
-  const dispatch = useAppDispatch()
-  useEffect(() => {
-    if (account) {
-      dispatch(fetchPoolsUserDataAsync(account))
-    }
-  }, [account, dispatch, fastRefresh])
-
-  const { pools, userDataLoaded } = useSelector((state: State) => ({
-    pools: state.pools.data,
-    userDataLoaded: state.pools.userDataLoaded,
-  }))
-  return { pools: pools.map(transformPool), userDataLoaded }
-}
-
-export const usePoolFromPid = (sousId: number): Pool => {
-  const pool = useSelector((state: State) => state.pools.data.find((p) => p.sousId === sousId))
-  return transformPool(pool)
-}
-
-export const useFetchCakeVault = () => {
-  const { account } = useWeb3React()
-  const { fastRefresh } = useRefresh()
-  const dispatch = useAppDispatch()
-
-  useEffect(() => {
-    dispatch(fetchCakeVaultPublicData())
-  }, [dispatch, fastRefresh])
-
-  useEffect(() => {
-    dispatch(fetchCakeVaultUserData({ account }))
-  }, [dispatch, fastRefresh, account])
-
-  useEffect(() => {
-    dispatch(fetchCakeVaultFees())
-  }, [dispatch])
-}
-
-export const useCakeVault = () => {
-  const {
-    totalShares: totalSharesAsString,
-    pricePerFullShare: pricePerFullShareAsString,
-    totalCakeInVault: totalCakeInVaultAsString,
-    estimatedCakeBountyReward: estimatedCakeBountyRewardAsString,
-    totalPendingCakeHarvest: totalPendingCakeHarvestAsString,
-    fees: { performanceFee, callFee, withdrawalFee, withdrawalFeePeriod },
-    userData: {
-      isLoading,
-      userShares: userSharesAsString,
-      cakeAtLastUserAction: cakeAtLastUserActionAsString,
-      lastDepositedTime,
-      lastUserActionTime,
-    },
-  } = useSelector((state: State) => state.pools.cakeVault)
-
-  const estimatedCakeBountyReward = useMemo(() => {
-    return new BigNumber(estimatedCakeBountyRewardAsString)
-  }, [estimatedCakeBountyRewardAsString])
-
-  const totalPendingCakeHarvest = useMemo(() => {
-    return new BigNumber(totalPendingCakeHarvestAsString)
-  }, [totalPendingCakeHarvestAsString])
-
-  const totalShares = useMemo(() => {
-    return new BigNumber(totalSharesAsString)
-  }, [totalSharesAsString])
-
-  const pricePerFullShare = useMemo(() => {
-    return new BigNumber(pricePerFullShareAsString)
-  }, [pricePerFullShareAsString])
-
-  const totalCakeInVault = useMemo(() => {
-    return new BigNumber(totalCakeInVaultAsString)
-  }, [totalCakeInVaultAsString])
-
-  const userShares = useMemo(() => {
-    return new BigNumber(userSharesAsString)
-  }, [userSharesAsString])
-
-  const cakeAtLastUserAction = useMemo(() => {
-    return new BigNumber(cakeAtLastUserActionAsString)
-  }, [cakeAtLastUserActionAsString])
-
-  return {
-    totalShares,
-    pricePerFullShare,
-    totalCakeInVault,
-    estimatedCakeBountyReward,
-    totalPendingCakeHarvest,
-    fees: {
-      performanceFee,
-      callFee,
-      withdrawalFee,
-      withdrawalFeePeriod,
-    },
-    userData: {
-      isLoading,
-      userShares,
-      cakeAtLastUserAction,
-      lastDepositedTime,
-      lastUserActionTime,
-    },
-  }
 }
 
 // Profile
