+++ src/views/Pools/components/CakeVaultCard/VaultCardActions/HasSharesActions.tsx
@@ -3,7 +3,8 @@ import { Flex, Text, IconButton, AddIcon, MinusIcon, useModal, Skeleton } from '
 import BigNumber from 'bignumber.js'
 import { getBalanceNumber } from 'utils/formatBalance'
 import { Pool } from 'state/types'
-import { useCakeVault, usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useCakeVault } from 'state/pools/hooks'
 import Balance from 'components/Balance'
 import NotEnoughTokensModal from '../../PoolCard/Modals/NotEnoughTokensModal'
 import { convertSharesToCake } from '../../../helpers'
